Refactored PieceColor into it's own file
